control-server: move to main
